requirements.txt
pymongo[tls, srv]
python-dotenv